combine add and edit controllers
